# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "001",
    "title": "Документация Fake-Post API",
    "description": "API для управления постами-> https://fake-posts-api.vercel.app."
  },
  "paths": {
    "/all-posts": {
      "get": {
        "summary": "Получить все посты",
        "description": "Возвращает список всех постов.",
        "responses": {
          "200": {
            "description": "Успешный ответ.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "title": {
                        "type": "string"
                      },
                      "content": {
                        "type": "string"
                      },
                      "imageUrl": {
                        "type": "string",
                        "example": "https://my-image/image.jpg"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/one-post/{id}": {
      "get": {
        "summary": "Получить пост по ID",
        "description": "Возвращает пост по заданному ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ с постом.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "imageUrl": {
                      "type": "string",
                      "example": "https://example.com/image.jpg"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка, если пост с таким ID не найден.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Нет такого ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/create": {
      "post": {
        "summary": "Создать новый пост",
        "description": "Создает новый пост",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пост успешно создан.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Успешно добавлено"
                    },
                    "newPost": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        },
                        "imageUrl": {
                          "type": "string",
                          "example": "https://example.com/image.jpg"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ошибка, если не предоставлены обязательные поля (title, content).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Ошибка при добавлении"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/delete/{id}": {
      "delete": {
        "summary": "Удалить пост по ID",
        "description": "Удаляет пост с заданным ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пост успешно удален.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Успешно удалено"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка, если пост с таким ID не найден.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Нет такого ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/update/{id}": {
      "put": {
        "summary": "Обновить пост по ID",
        "description": "Обновляет данные поста с заданным ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string",
                    "example": "https://updated-image.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пост успешно обновлен.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Успешно обновлено"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка, если пост с таким ID не найден.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Нет такого ID"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ошибка, если не предоставлены обязательные поля (title, content).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Ошибка при добавлении"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "description": "Dev",
      "url": "http://localhost:5001"
    },
    {
      "url": "https://fake-posts-api.vercel.app",
      "description": "Prod"
    }
  ]
}